Move virtualenv detection functions into their own module
