+++ src/pip/_internal/commands/configuration.py
@@ -6,9 +6,10 @@ from pip._internal.cli.base_command import Command
 from pip._internal.cli.status_codes import ERROR, SUCCESS
 from pip._internal.configuration import Configuration, kinds
 from pip._internal.exceptions import PipError
-from pip._internal.locations import running_under_virtualenv, site_config_file
+from pip._internal.locations import site_config_file
 from pip._internal.utils.deprecation import deprecated
 from pip._internal.utils.misc import get_prog
+from pip._internal.utils.virtualenv import running_under_virtualenv
 
 logger = logging.getLogger(__name__)
